APIMF-3324: Refactor, extracted common parse code
